# packages/api-contract/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InPulse API",
    "version": "1.0.0",
    "description": "由 packages/api-contract 的 Schema Registry 与 Route Registry 生成，禁止手工修改。"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/auth/mfa/enrollment/confirm": {
      "post": {
        "operationId": "confirmMfaEnrollment",
        "summary": "确认 TOTP 注册：条件接受 pending Secret 的当前 time-step，启用因子、签发恢复码，并在同一事务撤销受限 Session、轮换为完整 Session/CSRF。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmMfaEnrollmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录或 Session 失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但无全局权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本冲突、重复操作或状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "字段校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-inpulse-policy": {
          "authPolicy": "session",
          "csrfPolicy": "required",
          "idempotencyPolicy": "securityFlow",
          "idempotencyExceptionAdr": "ADR-023",
          "idempotencyContractVersion": "none",
          "idempotencyFingerprintVersion": "none",
          "behaviorHeaders": "none",
          "idempotencyReplayPolicy": "none",
          "replayAuthorizationPolicy": "none",
          "securityFlowPolicy": {
            "singleConsumptionOrNaturalIdempotency": "条件激活对应 ENROLLING generation；恢复码 Hash 一次签发、原子失效旧批次，旧 Session 撤销并轮换新 Session/CSRF",
            "clientRecoveryPath": "清 Cookie 后重新登录；若因子已激活，完成 MFA 后于下一 time-step 重认证并轮换恢复码"
          },
          "versionPolicy": "none",
          "concurrencyPolicy": "none",
          "auditAction": "none"
        },
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMfaEnrollmentRequest"
              }
            }
          }
        }
      }
    },
    "/auth/mfa/enrollment/start": {
      "post": {
        "operationId": "startMfaEnrollment",
        "summary": "管理员首次注册 TOTP：按 user → factor → current Session 锁序条件创建新 pending，返回只展示一次 Secret 与 otpauth URI。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartMfaEnrollmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录或 Session 失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "已登录但无全局权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本冲突、重复操作或状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "字段校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-inpulse-policy": {
          "authPolicy": "session",
          "csrfPolicy": "required",
          "idempotencyPolicy": "securityFlow",
          "idempotencyExceptionAdr": "ADR-023",
          "idempotencyContractVersion": "none",
          "idempotencyFingerprintVersion": "none",
          "behaviorHeaders": "none",
          "idempotencyReplayPolicy": "none",
          "replayAuthorizationPolicy": "none",
          "securityFlowPolicy": {
            "singleConsumptionOrNaturalIdempotency": "同一旧 enrollment generation 只允许一个请求创建新 pending；按 user → factor → current Session 锁序条件更新，失败方返回 409",
            "clientRecoveryPath": "重新登录读取当前 enrollment generation，再以该 generation 重新开始"
          },
          "versionPolicy": "none",
          "concurrencyPolicy": "none",
          "auditAction": "none"
        },
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMfaEnrollmentRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmMfaEnrollmentRequest": {
        "type": "object",
        "properties": {
          "expectedEnrollmentGeneration": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "expectedEnrollmentGeneration",
          "code"
        ],
        "additionalProperties": false
      },
      "ConfirmMfaEnrollmentResponse": {
        "type": "object",
        "properties": {
          "csrfToken": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "authState": {
            "type": "string",
            "const": "AUTHENTICATED"
          },
          "recoveryCodes": {
            "minItems": 10,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-HJ-NP-Z2-9]{20}$"
            }
          }
        },
        "required": [
          "csrfToken",
          "authState",
          "recoveryCodes"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "requestId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "StartMfaEnrollmentRequest": {
        "type": "object",
        "properties": {
          "expectedEnrollmentGeneration": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "expectedEnrollmentGeneration"
        ],
        "additionalProperties": false
      },
      "StartMfaEnrollmentResponse": {
        "type": "object",
        "properties": {
          "enrollmentGeneration": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "secret": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[A-Z2-7]{32}$"
          },
          "otpauthUri": {
            "type": "string",
            "maxLength": 2048,
            "pattern": "^otpauth:\\/\\/totp\\/.*"
          }
        },
        "required": [
          "enrollmentGeneration",
          "secret",
          "otpauthUri"
        ],
        "additionalProperties": false
      },
...
    }
  }
}